Add ability to add items to db, plus a few small fixes
Form POST's data to backend for processing and is added to db as a new
document.
Small fix to comment handling, whereby if a comment was posted to a
newly added drink, it would crash.
Changed comment section of drink page to encourage user to sign in/up to
make a comment if no user logged in.