Refactor code for flashed messages & add / edit drink form
Removed individual code for flashed messages on each page and moved to
it's own template and called from base.html so it covers all pages.

HTML for the drinks form for both adding and editing a drink was
similar. Refactored the Jinja2 elements of the pages so the same code
could be used to generate the form for both adding and editing a drink.