Use localStorage to remember some UI settings
Settings for the open/close state of the tabs on index and also the
name of the last selected tab are stored in localStorage so when a user
returns to the page, it looks the same. It also ensures consistency
across all pages if these tabs are reused anywhere else.