Add .js to fix card heights
When viewing the categories using the category tab filter on the index
page, depending on the amount of text to display, the height of the
cards could vary. This small bit of js ensures on each page load, the
card heights are all set the same, to the value of the highest card.

Could not use css as it would only make the outer divs equal height,
not the card itself so it still looked 'wonky'.