Fix issue with drink deletion
When drinks were being deleted, if a user had favorited a drink, upon
deletion the drink was still stored as a favorite for the user.
Add fix to delete_drink() so that on deletion of a drink, the code
gets the names of all users who had favorited that drink, and removes
the drink_id from the users favoritesTxt array.