Change how category boxes are displayed on homepage
this was causing issues on page load if boxes were hidden - it was
setting all heights to 0. Have removed js and now allowing heights to
resize dynamically based on content for widths > 600px.

On views <= 600px media queries used to set a fixed height, and in some
cases reduced font-size, to ensure all lengths of text fit comfortably
within it's container.

Not a perfect solution, but acceptable for now.