Fix bug with displaying user submitted drink stats
Issue identified whereby if a user has submitted only a few drinks, and
none of them had recieved any views, comments, or favorites, even with
a value of zero, it would still show as the users most viewed (etc)
drink. Fixed by adding a check to account() in app.py to return None
for any drinks whose value for views (etc) is 0.

Also encountered a bug where if there are less than 4 drinks in the db
that were submitted by users other than the 1 logged in, the function
for get_suggestions() would fail as it would attempt to store 4 drinks,
but run out before this. Unlikely any users would encounter this bug as
the db is populated with enough drinks, but fixed anyway by only
executing if 4+ drinks available to choose from.