Fix bug with view counter
View counter not updating correctly. Does not inc when viewing a drink
where the referring path is '/all_drinks'. This started when the route
was renamed to allow the landing page route to be '/'. The view counter
excludes certain paths to stop the views increasing when they
shouldn't, for example adding / removing favorites. Because one of the
paths excluded was the string 'drink', this had the side effect of
blocking the view counter incrementing when coming from the index route
of 'all_drinks'. To fix, changed string 'drink' to 'drink/' in the
exlude paths list.