Add checks to stop users doing things they shouldn't
Checks added to each route to stop users and guests either accessing
pages they aren't allowed to (eg other users account page), or
performing actions they are not allowed to (eg delete another users
drink). Where a potential violation is discovered, the user is
redirected back to another page and a message is flashed to inform the
user they cannot do that.