Change page load behaviour on search
When a user searches, this update will make the page jump to the search
results and not just stay at the top. On some displays, when search
results load, it can look as though nothing has happened as when
viewing the top of the page, a user is still presented with search
options, and not the actual results.